requirements.txt
numpy>=1.20.0
matplotlib>=3.3.0
pyserial>=3.5
pyyaml>=5.4
scipy>=1.7.0
pytest>=6.2.0
pytest-cov>=2.12.0